+++ src/pages/ListPage/MonthlyChart/styles.module.scss
@@ -24,7 +24,7 @@
   height: 32px;
   margin-bottom: 24px;
 
-  h1 {
+  h2 {
     @include font-weight-size(700, 24px);
     line-height: 26px;
     color: white;
